+++ drivers/mfd/Kconfig
@@ -27,5 +27,6 @@ source "drivers/mfd/Kconfig.nct38xx"
 source "drivers/mfd/Kconfig.npm1300"
 source "drivers/mfd/Kconfig.npm6001"
 source "drivers/mfd/Kconfig.lpflexcomm"
+source "drivers/mfd/Kconfig.tle9104"
 
 endif # MFD
