+++ dts/bindings/gpio/infineon,tle9104-gpio.yaml
@@ -1,0 +1,27 @@
+#
+# Copyright (c) 2024 SILA Embedded Solutions GmbH
+#
+# SPDX-License-Identifier: Apache-2.0
+#
+
+description: Infineon TLE9104 4-channel powertrain switch GPIO instance
+
+compatible: "infineon,tle9104-gpio"
+
+include: gpio-controller.yaml
+
+on-bus: tle9104
+
+properties:
+  "#gpio-cells":
+    const: 2
+
+  ngpios:
+    type: int
+    required: true
+    const: 4
+    description: Number of GPIOs supported
+
+gpio-cells:
+  - pin
+  - flags
