--- dts/bindings/gpio/infineon,tle9104.yaml
@@ -1,49 +1,0 @@
-#
-# Copyright (c) 2023 SILA Embedded Solutions GmbH
-#
-# SPDX-License-Identifier: Apache-2.0
-#
-
-description: Infineon TLE9104 4-channel powertrain switch
-
-compatible: "infineon,tle9104"
-
-include: [gpio-controller.yaml, spi-device.yaml]
-
-properties:
-  "#gpio-cells":
-    const: 2
-
-  ngpios:
-    type: int
-    required: true
-    const: 4
-    description: Number of GPIOs supported
-
-  en-gpios:
-    type: phandle-array
-    description: "GPIO for enable"
-
-  resn-gpios:
-    type: phandle-array
-    description: "GPIO for reset"
-
-  in1-gpios:
-    type: phandle-array
-    description: "GPIO for controlling OUT1"
-
-  in2-gpios:
-    type: phandle-array
-    description: "GPIO for controlling OUT2"
-
-  in3-gpios:
-    type: phandle-array
-    description: "GPIO for controlling OUT3"
-
-  in4-gpios:
-    type: phandle-array
-    description: "GPIO for controlling OUT4"
-
-gpio-cells:
-  - pin
-  - flags
