+++ tests/drivers/build_all/gpio/app.overlay
@@ -299,13 +299,19 @@
 
 			test_spi_tle9104: tle9104@4 {
 				compatible = "infineon,tle9104";
+				status = "okay";
 				spi-max-frequency = <0>;
 				reg = <0x04>;
-				gpio-controller;
-				#gpio-cells = <2>;
-				ngpios = <4>;
 				resn-gpios = <&test_gpio 0 0>;
 				en-gpios = <&test_gpio 0 0>;
+
+				tle9104_gpio: tle9104_gpio {
+					compatible = "infineon,tle9104-gpio";
+					status = "okay";
+					gpio-controller;
+					#gpio-cells = <2>;
+					ngpios = <4>;
+				};
 			};
 		};
 	};
