+++ tests/drivers/build_all/gpio/prj.conf
@@ -4,4 +4,7 @@ CONFIG_TEST_USERSPACE=y
 CONFIG_I2C=y
 CONFIG_GPIO_PCA95XX_INTERRUPT=y
 CONFIG_SPI=y
+CONFIG_MFD=y
+CONFIG_MFD_INIT_PRIORITY=80
+CONFIG_GPIO_TLE9104_INIT_PRIORITY=81
 CONFIG_GPIO_BD8LB600FS_INIT_PRIORITY=81
